Extracting the test framework in a separate include file. Refactoring tests to avoid code duplication.
