+++ MQTTPacket/test/test1.c
@@ -14,235 +14,12 @@
  *    Ian Craggs - initial API and implementation and/or initial documentation
  *******************************************************************************/
 
+/***
+  Tests for MQTTV3 serialization and deserialization
+***/
 
+#include "test_framework.h"
 #include "MQTTPacket.h"
-#include <string.h>
-#include <stdlib.h>
-#include <stdio.h>
-
-#if !defined(_WINDOWS)
-	#include <sys/time.h>
-  	#include <sys/socket.h>
-	#include <unistd.h>
-  	#include <errno.h>
-#else
-#include <winsock2.h>
-#include <ws2tcpip.h>
-#define MAXHOSTNAMELEN 256
-#define EAGAIN WSAEWOULDBLOCK
-#define EINTR WSAEINTR
-#define EINPROGRESS WSAEINPROGRESS
-#define EWOULDBLOCK WSAEWOULDBLOCK
-#define ENOTCONN WSAENOTCONN
-#define ECONNRESET WSAECONNRESET
-#endif
-
-#define ARRAY_SIZE(a) (sizeof(a) / sizeof(a[0]))
-
-struct Options
-{
-	char* connection;         /**< connection to system under test. */
-	char** haconnections;
-	int hacount;
-	int verbose;
-	int test_no;
-} options =
-{
-	"tcp://localhost:1883",
-	NULL,
-	0,
-	0,
-	0,
-};
-
-void usage()
-{
-
-}
-
-void getopts(int argc, char** argv)
-{
-	int count = 1;
-
-	while (count < argc)
-	{
-		if (strcmp(argv[count], "--test_no") == 0)
-		{
-			if (++count < argc)
-				options.test_no = atoi(argv[count]);
-			else
-				usage();
-		}
-		else if (strcmp(argv[count], "--connection") == 0)
-		{
-			if (++count < argc)
-			{
-				options.connection = argv[count];
-				printf("\nSetting connection to %s\n", options.connection);
-			}
-			else
-				usage();
-		}
-		else if (strcmp(argv[count], "--haconnections") == 0)
-		{
-			if (++count < argc)
-			{
-				char* tok = strtok(argv[count], " ");
-				options.hacount = 0;
-				options.haconnections = malloc(sizeof(char*) * 5);
-				while (tok)
-				{
-					options.haconnections[options.hacount] = malloc(strlen(tok) + 1);
-					strcpy(options.haconnections[options.hacount], tok);
-					options.hacount++;
-					tok = strtok(NULL, " ");
-				}
-			}
-			else
-				usage();
-		}
-		else if (strcmp(argv[count], "--verbose") == 0)
-		{
-			options.verbose = 1;
-			printf("\nSetting verbose on\n");
-		}
-		count++;
-	}
-}
-
-
-#define LOGA_DEBUG 0
-#define LOGA_INFO 1
-#include <stdarg.h>
-#include <time.h>
-#include <sys/timeb.h>
-void MyLog(int LOGA_level, char* format, ...)
-{
-	static char msg_buf[256];
-	va_list args;
-	struct timeb ts;
-
-	struct tm *timeinfo;
-
-	if (LOGA_level == LOGA_DEBUG && options.verbose == 0)
-	  return;
-
-	ftime(&ts);
-	timeinfo = localtime(&ts.time);
-	strftime(msg_buf, 80, "%Y%m%d %H%M%S", timeinfo);
-
-	sprintf(&msg_buf[strlen(msg_buf)], ".%.3hu ", ts.millitm);
-
-	va_start(args, format);
-	vsnprintf(&msg_buf[strlen(msg_buf)], sizeof(msg_buf) - strlen(msg_buf), format, args);
-	va_end(args);
-
-	printf("%s\n", msg_buf);
-	fflush(stdout);
-}
-
-
-#if defined(WIN32) || defined(_WINDOWS)
-#define mqsleep(A) Sleep(1000*A)
-#define START_TIME_TYPE DWORD
-static DWORD start_time = 0;
-START_TIME_TYPE start_clock(void)
-{
-	return GetTickCount();
-}
-#elif defined(AIX)
-#define mqsleep sleep
-#define START_TIME_TYPE struct timespec
-START_TIME_TYPE start_clock(void)
-{
-	static struct timespec start;
-	clock_gettime(CLOCK_REALTIME, &start);
-	return start;
-}
-#else
-#define mqsleep sleep
-#define START_TIME_TYPE struct timeval
-START_TIME_TYPE start_clock(void)
-{
-	struct timeval start_time;
-	gettimeofday(&start_time, NULL);
-	return start_time;
-}
-#endif
-
-
-#if defined(WIN32)
-long elapsed(START_TIME_TYPE start_time)
-{
-	return GetTickCount() - start_time;
-}
-#elif defined(AIX)
-#define assert(a)
-long elapsed(struct timespec start)
-{
-	struct timespec now, res;
-
-	clock_gettime(CLOCK_REALTIME, &now);
-	ntimersub(now, start, res);
-	return (res.tv_sec)*1000L + (res.tv_nsec)/1000000L;
-}
-#else
-long elapsed(START_TIME_TYPE start_time)
-{
-	struct timeval now, res;
-
-	gettimeofday(&now, NULL);
-	timersub(&now, &start_time, &res);
-	return (res.tv_sec)*1000 + (res.tv_usec)/1000;
-}
-#endif
-
-
-#define assert(a, b, c, d) myassert(__FILE__, __LINE__, a, b, c, d)
-#define assert1(a, b, c, d, e) myassert(__FILE__, __LINE__, a, b, c, d, e)
-
-int tests = 0;
-int failures = 0;
-FILE* xml;
-START_TIME_TYPE global_start_time;
-char output[3000];
-char* cur_output = output;
-
-
-void write_test_result()
-{
-	long duration = elapsed(global_start_time);
-
-	fprintf(xml, " time=\"%ld.%.3ld\" >\n", duration / 1000, duration % 1000);
-	if (cur_output != output)
-	{
-		fprintf(xml, "%s", output);
-		cur_output = output;
-	}
-	fprintf(xml, "</testcase>\n");
-}
-
-
-void myassert(char* filename, int lineno, char* description, int value, char* format, ...)
-{
-	++tests;
-	if (!value)
-	{
-		va_list args;
-
-		++failures;
-		printf("Assertion failed, file %s, line %d, description: %s\n", filename, lineno, description);
-
-		va_start(args, format);
-		vprintf(format, args);
-		va_end(args);
-
-		cur_output += sprintf(cur_output, "<failure type=\"%s\">file %s, line %d </failure>\n",
-                        description, filename, lineno);
-	}
-    else
-    	MyLog(LOGA_DEBUG, "Assertion succeeded, file %s, line %d, description: %s", filename, lineno, description);
-}
 
 #define min(a, b) ((a < b) ? a : b)
